Mutations directly accessible, not behind user resolver
